+++ Xbim.Ifc2x3/Interfaces/IFC4/IfcColourRgb.cs
@@ -20,21 +20,21 @@ namespace Xbim.Ifc2x3.PresentationResource
 		{ 
 			get
 			{
-				return new Ifc4.MeasureResource.IfcNormalisedRatioMeasure((double)Red);
+				return new Ifc4.MeasureResource.IfcNormalisedRatioMeasure(Red);
 			} 
 		}
 		Ifc4.MeasureResource.IfcNormalisedRatioMeasure IIfcColourRgb.Green 
 		{ 
 			get
 			{
-				return new Ifc4.MeasureResource.IfcNormalisedRatioMeasure((double)Green);
+				return new Ifc4.MeasureResource.IfcNormalisedRatioMeasure(Green);
 			} 
 		}
 		Ifc4.MeasureResource.IfcNormalisedRatioMeasure IIfcColourRgb.Blue 
 		{ 
 			get
 			{
-				return new Ifc4.MeasureResource.IfcNormalisedRatioMeasure((double)Blue);
+				return new Ifc4.MeasureResource.IfcNormalisedRatioMeasure(Blue);
 			} 
 		}
 	}
