+++ Xbim.Ifc2x3/Interfaces/IFC4/IfcEllipse.cs
@@ -20,14 +20,14 @@ namespace Xbim.Ifc2x3.GeometryResource
 		{ 
 			get
 			{
-				return new Ifc4.MeasureResource.IfcPositiveLengthMeasure((double)SemiAxis1);
+				return new Ifc4.MeasureResource.IfcPositiveLengthMeasure(SemiAxis1);
 			} 
 		}
 		Ifc4.MeasureResource.IfcPositiveLengthMeasure IIfcEllipse.SemiAxis2 
 		{ 
 			get
 			{
-				return new Ifc4.MeasureResource.IfcPositiveLengthMeasure((double)SemiAxis2);
+				return new Ifc4.MeasureResource.IfcPositiveLengthMeasure(SemiAxis2);
 			} 
 		}
 	}
