+++ Xbim.Ifc2x3/Interfaces/IFC4/IfcRoundedRectangleProfileDef.cs
@@ -20,7 +20,7 @@ namespace Xbim.Ifc2x3.ProfileResource
 		{ 
 			get
 			{
-				return new Ifc4.MeasureResource.IfcPositiveLengthMeasure((double)RoundingRadius);
+				return new Ifc4.MeasureResource.IfcPositiveLengthMeasure(RoundingRadius);
 			} 
 		}
 	}
